% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-default.R
\name{geoarrow_infer_schema}
\alias{geoarrow_infer_schema}
\title{Infer a GeoArrow-native type from a vector}
\usage{
geoarrow_infer_schema(x, ..., promote_multi = TRUE, coord_type = NULL)
}
\arguments{
\item{promote_multi}{Use \code{TRUE} to return a MULTI type when both normal and
MULTI elements are in the same array.}

\item{coord_type}{Specify the coordinate type to use if returning}

\item{handleable}{An object implementing \code{\link[=wk_vector_meta]{wk_vector_meta()}} and \code{\link[=wk_meta]{wk_meta()}}.}
}
\value{
A \link[=as_nanoarrow_schema]{nanoarrow_schema}
}
\description{
Infer a GeoArrow-native type from a vector
}
\examples{
geoarrow_infer_schema(wk::wkt("POINT (0 1)"))

}
